% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scale.R
\name{compute.scale}
\alias{compute.scale}
\title{Calcola una scala a partire dagli items}
\usage{
compute.scale(items, prop.na = 0.2, na.action = "rowMeans",
  out.names = "out", params = list(invert = FALSE, values.ranges = NULL,
  transf = function(items) rowMeans(items)))
}
\arguments{
\item{items}{a matrix o data.frame}

\item{prop.na}{proportion of NAs allowed before droping the observation. .20 by default}

\item{na.action}{how to deal NAs. "rowMeans" by default.}

\item{out.names}{a vector of colnames to be used in the output. "out" by default.}
}
\description{
Calcola una scala a partire dagli items
}
\details{
Calcola una scala a partire dagli items
}
\examples{
DAS=compute.scale(items_DAS,prop.na=.20,
                  params = scale_das_params)
                  
                  summary(DAS)
                  pairsSigmaPsi(DAS)
}
\author{
Livio Finos
}

