% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items_DAS.R
\docType{data}
\name{items_DAS}
\alias{items_DAS}
\title{Sample data set of DAS items}
\format{a data.frame of 50 rows and 32 items}
\source{
an artificial data set to show the format to be used in compute.scale() function
}
\usage{
items_DAS
}
\description{
Sample data set of DAS items
}
\author{
livio finos, Gianmarco Altoe'
}
\seealso{
compute.scale
}

