% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_sigmapsi.R
\name{pairsSigmaPsi}
\alias{pairsSigmaPsi}
\title{A customization of pairs function}
\usage{
pairsSigmaPsi(X, ...)
}
\arguments{
\item{X}{a matrix or data.frame}
}
\description{
A customization of pairs function
}
\examples{
pairsSigmaPsi(matrix(rnorm(300),100,3))
}
\author{
Gianmarco Altoè, Laura Boscaro, Marta Bagno
}

