% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_var-cov.R
\name{panel.hist}
\alias{panel.hist}
\title{Funzioni utili per pairs.sigmapsy}
\usage{
panel.hist(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{an object}

\item{y}{an object}

\item{digits}{2}

\item{prefix}{""}

\item{cex.cor}{an object}
}
\description{
Funzioni utili per pairs.sigmapsy
}
\author{
Laura Boscaro, Marta Bagno
}

