% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXresid.R
\name{plotXresid}
\alias{plotXresid}
\title{plotXresid}
\usage{
plotXresid(mod, x)
}
\arguments{
\item{mod}{lm model}

\item{x}{name of the term to be ploted}
}
\description{
plot x vs residual of an lm model
}

