% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowSums.bylist}
\alias{rowSums.bylist}
\title{rowSums and rowMeans performed for each element of a list and with a slight modification}
\usage{
rowSums.bylist(items, ids)
}
\value{
a data.frame of dimension ncol(items)Xlength(ids)
}
\description{
rowSums and rowMeans with a slight modification
}
\author{
Livio Finos
}

