% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scale_params.R
\docType{data}
\name{scale_params}
\alias{scale_biaq_params}
\alias{scale_das_params}
\alias{scale_ecr_params}
\alias{scale_params}
\alias{scale_scl_params}
\alias{scale_tas20_params}
\alias{scale_tas24_params}
\title{Setting for computing scales of a few tests}
\format{An object of class \code{list} of length 4.}
\usage{
scale_ecr_params
}
\description{
DAS
item dal 1 al 15, 18 e 19: i valori degli item vanno inseriti da 5 (tot d’accordo, sempre) a 0 (tot disaccordo, mai)
item dal 16, 17, 20 al 22, dal 25 al 28: i valori degli item vanno inseriti da 0 (sempre, mai) a 5 (mai, sempre) 
item 23, 24: da 4 a 0
item 29, 30: 0 – 1
item 31: 0 (estremamente infelice) a 6 (perfetto)
item 32: a=5, b=4, c=3, d=2, e=1, f=0

calcola F1. dyadic consensus: somma item: 1+2+3+5+7+8+9+10+11+12+13+14+15
calcola F2. dyadic satisfaction: somma item: 16+17++18+19+20+21+22+23+31+32
calcola F3: affectional expression: somma item: 4+6+29+30
calcola F4: diadi cohesion: somma item: 24+25+26+27+28
totale DAS = f1+f2+f3+f4

valori normativi in punti T (media 50, ds 10 - norma 40-60), separati per sposati e divorziati.


ECR 
(36 item, da 7 a 1 - massimo 252 –tot disaccordo-, minimo 36 – totale accordo-)
girare item (mettergli r): 3, 15, 19, 22,25, 27, 29, 31, 33, 35.
item dispari = evitamento (18 item: min 18 relaz ok, max 126 relaz ko)
item pari = ansietà (18 item: min 18 ok, max 126 relaz ko)
misura qualitativa di profilo, per soggetto. Non ci sono cut off. La media per scala è 63, ds 12,6
trasforma in punti T: 50:63=10:12,6
cut off sensato: 
 norma 50-76
76 a 82 border problematico
>82 patologico evitamento o ansietà


  fattori TAS-20 e globale 
  fattori TAS-24 e globale
  f1 difficoltà identificare sentimenti
  f2 difficoltà comprendere sentimenti
  f3 pensiero orient esternamente
  cut off globale: norma punteggi inferiori a 55 ? soggetti non alessitimici; 
  punteggi compresi tra 55 e 64 ? soggetti borderline; 
  punteggi superiori a 64   ? soggetti alessitimici.
}
\details{
A series of preformed setting for computing scales of a few tests
}
\keyword{datasets}

