% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmaPsi.R
\docType{package}
\name{SigmaPsi}
\alias{SigmaPsi}
\alias{SigmaPsi-package}
\title{SigmaPsi}
\description{
an R package for Psychometrics
}
\author{
Author: Livio Finos, Gianmarco Altoè, Massimiliano Pastore
}

